+++ database/migrations/2025_06_06_190200_create_movies_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->integer('duration');
+            $table->unsignedSmallInteger('duration');
             $table->string('poster_url')->nullable();
             $table->timestamps();
         });
